+++ simvue/__init__.py
@@ -1,4 +1,5 @@
 from simvue.run import Run
 from simvue.client import Client
 from simvue.handler import Handler
-__version__ = '0.7.0'
+from simvue.models import RunInput
+__version__ = '0.11.1'