+++ simvue/api.py
@@ -1,3 +1,5 @@
+import copy
+import json
 import requests
 from tenacity import retry, wait_exponential, stop_after_attempt
 
@@ -7,6 +9,15 @@ RETRY_MIN = 4
 RETRY_MAX = 10
 RETRY_STOP = 5
 
+def set_json_header(headers):
+    """
+    Return a copy of the headers with Content-Type set to
+    application/json
+    """
+    headers = copy.deepcopy(headers)
+    headers['Content-Type'] = 'application/json'
+    return headers
+
 @retry(wait=wait_exponential(multiplier=RETRY_MULTIPLIER, min=RETRY_MIN, max=RETRY_MAX),
                              stop=stop_after_attempt(RETRY_STOP))
 def post(url, headers, data, is_json=True):
@@ -14,9 +25,11 @@ def post(url, headers, data, is_json=True):
     HTTP POST with retries
     """
     if is_json:
-        response = requests.post(url, headers=headers, json=data, timeout=DEFAULT_API_TIMEOUT)
-    else:
-        response = requests.post(url, headers=headers, data=data, timeout=DEFAULT_API_TIMEOUT)
+        data = json.dumps(data)
+        headers = set_json_header(headers)
+    response = requests.post(url, headers=headers, data=data, timeout=DEFAULT_API_TIMEOUT)
+    if response.status_code not in (200, 201, 409):
+        raise Exception('HTTP error')
 
     return response
 
@@ -27,8 +40,9 @@ def put(url, headers, data, is_json=True, timeout=DEFAULT_API_TIMEOUT):
     HTTP PUT with retries
     """
     if is_json:
-        response = requests.put(url, headers=headers, json=data, timeout=timeout)
-    else:
-        response = requests.put(url, headers=headers, data=data, timeout=timeout)
+        data = json.dumps(data)
+        headers = set_json_header(headers)
+    response = requests.put(url, headers=headers, data=data, timeout=timeout)
+    response.raise_for_status()
 
     return response